backend/requirements.txt
Django==4.2.13
djangorestframework==3.15.1
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
Pillow==10.3.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
django-filter==24.2
gunicorn==22.0.0
